Generalize ChemCheck/ConcordCheck to IdentifierCheck/ConcordCheck in shared __init__.py
ChemCheck lacked a `compendium` field, so test functions hardcoded
outputs["chemicals"]. Moving both types to tests/pipeline/checks/__init__.py
with a `compendium` field on IdentifierCheck lets future per-compendium
check files (test_diseases.py, test_proteins.py, …) reuse the same types
without duplication. No logic changes — purely a rename and extraction.
