+++ tests/pipeline/checks/test_chemicals.py
@@ -25,69 +25,56 @@ appropriate session fixture name.
 Run:
     uv run pytest tests/pipeline/checks/test_chemicals.py --pipeline --no-cov -v
 """
-from typing import NamedTuple
-
 import pytest
 
 from src.categories import CHEMICAL_ENTITY
+from tests.pipeline.checks import ConcordCheck, IdentifierCheck
 from tests.pipeline.conftest import _any_concord_xrefs, get_curies_and_types_from_ids_file
 
 # ---------------------------------------------------------------------------
-# ID-presence check type and tables
+# ID-presence check tables
 # ---------------------------------------------------------------------------
 
 
-class ChemCheck(NamedTuple):
-    fixture: str        # session fixture name, e.g. "mesh_pipeline_outputs"
-    curie: str          # CURIE to look for in the chemicals ID file
-    expected_type: str  # Biolink type (used in failure messages; also asserted if the
-                        # intermediate file includes a type column, e.g. for UMLS)
-    issue: str          # GitHub issue URL that motivated this check
-
-
-EXPECTED_IN_CHEMICALS: list[ChemCheck] = [
-    ChemCheck(
+EXPECTED_IN_CHEMICALS: list[IdentifierCheck] = [
+    IdentifierCheck(
         "mesh_pipeline_outputs",
+        "chemicals",
         "MESH:C000598555",
         CHEMICAL_ENTITY,
         "https://github.com/NCATSTranslator/Babel/issues/708",
     ),
-    ChemCheck(
+    IdentifierCheck(
         "mesh_pipeline_outputs",
+        "chemicals",
         "MESH:C100843",
         CHEMICAL_ENTITY,
         "https://github.com/NCATSTranslator/Babel/issues/708",
     ),
     # D08.211 Coenzymes — non-protein small molecules that were previously excluded from
     # chemicals by a blanket D08 exclusion.  They should be CHEMICAL_ENTITY.
-    ChemCheck(
+    IdentifierCheck(
         "mesh_pipeline_outputs",
+        "chemicals",
         "MESH:D009243",  # Nicotinamide Adenine Dinucleotide (NAD) — D08.211.060
         CHEMICAL_ENTITY,
         "https://github.com/NCATSTranslator/Babel/issues/675",
     ),
-    ChemCheck(
+    IdentifierCheck(
         "mesh_pipeline_outputs",
+        "chemicals",
         "MESH:D003067",  # Coenzyme A — D08.211.190
         CHEMICAL_ENTITY,
         "https://github.com/NCATSTranslator/Babel/issues/675",
     ),
 ]
 
-NOT_IN_CHEMICALS: list[ChemCheck] = []
+NOT_IN_CHEMICALS: list[IdentifierCheck] = []
 
 
 # ---------------------------------------------------------------------------
-# Direct-xref check type and tables
+# Direct-xref check tables
 # ---------------------------------------------------------------------------
-
-
-class ConcordCheck(NamedTuple):
-    fixture: str        # session fixture providing the concords directory path
-    curie1: str
-    curie2: str
-    should_xref: bool   # True = must be a direct xref pair; False = must NOT be
-    issue: str          # GitHub issue URL that motivated this check
 
 
 EXPECTED_XREF: list[ConcordCheck] = []
@@ -110,10 +97,10 @@ EXPECTED_NO_XREF: list[ConcordCheck] = [
 
 @pytest.mark.pipeline
 @pytest.mark.parametrize("check", EXPECTED_IN_CHEMICALS, ids=[c.curie for c in EXPECTED_IN_CHEMICALS])
-def test_curie_in_chemicals(request, check: ChemCheck) -> None:
+def test_curie_in_chemicals(request, check: IdentifierCheck) -> None:
     """CURIE must appear in the chemicals intermediate ID file."""
     outputs = request.getfixturevalue(check.fixture)
-    ids = get_curies_and_types_from_ids_file(outputs["chemicals"])
+    ids = get_curies_and_types_from_ids_file(outputs[check.compendium])
     assert check.curie in ids, (
         f"{check.curie} not found in chemicals "
         f"(expected type {check.expected_type}; see {check.issue})"
@@ -128,10 +115,10 @@ def test_curie_in_chemicals(request, check: ChemCheck) -> None:
 
 @pytest.mark.pipeline
 @pytest.mark.parametrize("check", NOT_IN_CHEMICALS, ids=[c.curie for c in NOT_IN_CHEMICALS] or None)
-def test_curie_not_in_chemicals(request, check: ChemCheck) -> None:
+def test_curie_not_in_chemicals(request, check: IdentifierCheck) -> None:
     """CURIE must NOT appear in the chemicals intermediate ID file."""
     outputs = request.getfixturevalue(check.fixture)
-    ids = get_curies_and_types_from_ids_file(outputs["chemicals"])
+    ids = get_curies_and_types_from_ids_file(outputs[check.compendium])
     assert check.curie not in ids, (
         f"{check.curie} found in chemicals but should not be "
         f"(expected type {check.expected_type} elsewhere; see {check.issue})"
